+++ tests/test_basemodel.py
@@ -115,6 +115,7 @@ def test_camelize():
     assert dump == {
         "comment": None,
         "grants": None,
+        "individualGrants": None,
         "name": "my_schema",
         "storageRoot": None,
         "tables": [
@@ -139,6 +140,7 @@ def test_camelize():
                 ],
                 "comment": None,
                 "grants": None,
+                "individualGrants": None,
                 "name": "AAPL",
                 "properties": None,
                 "catalogName": "my_catalog",
@@ -171,6 +173,7 @@ def test_camelize():
                 ],
                 "comment": None,
                 "grants": None,
+                "individualGrants": None,
                 "name": "GOOGL",
                 "properties": None,
                 "catalogName": "my_catalog",